pyproject.toml
[tool.poetry]
name = "memegen"
version = "0.1.0"
description = "Automatic memes generation bot"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
aiohttp-socks = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
